+++ browser-app/package.json
@@ -58,7 +58,7 @@
     "@theia/variable-resolver": "1.55.0",
     "@theia/vsx-registry": "1.55.0",
     "@theia/workspace": "1.55.0",
-    "theia-ai-exerciser": "0.0.0"
+    "exerciser": "0.0.0"
   },
   "devDependencies": {
     "@theia/cli": "1.55.0"
