+++ electron-app/package.json
@@ -16,7 +16,7 @@
     "@theia/process": "1.55.0",
     "@theia/terminal": "1.55.0",
     "@theia/workspace": "1.55.0",
-    "theia-ai-exerciser": "0.0.0"
+    "exerciser": "0.0.0"
   },
   "devDependencies": {
     "@theia/cli": "1.55.0",
