+++ exerciser/README.md
