+++ exerciser/src/browser/utils/tool-functions/function-names.ts
@@ -1,0 +1,4 @@
+export const CREATE_FILE_FUNCTION_ID = 'createFile';
+export const GET_FILE_CONTENT_FUNCTION_ID = 'getFileContent';
+export const GET_WORKSPACE_FILES_FUNCTION_ID = 'getWorkspaceFileList';
+export const COMPARE_FILES_CONTENT_FUNCTION_ID = 'compareFilesContent'
