+++ package.json
@@ -18,6 +18,6 @@
     "lerna": "2.4.0"
   },
   "workspaces": [
-    "theia-ai-exerciser", "browser-app", "electron-app"
+    "exerciser", "browser-app", "electron-app"
   ]
 }
