--- theia-ai-exerciser/src/browser/coding-exercise-agent-frontend-module.ts
@@ -1,26 +1,0 @@
-/**
- * Generated using theia-extension-generator
- */
-
-import { ContainerModule } from 'inversify';
-import { CodingExerciseAgent } from './coding-exercise-agent';
-//import { FileService } from '@theia/filesystem/lib/browser/file-service';
-//import { WorkspaceService } from '@theia/workspace/lib/browser';
-//import { TerminalService } from '@theia/terminal/lib/browser/base/terminal-service';
-import { ChatAgent } from '@theia/ai-chat/lib/common';
-import { Agent, ToolProvider } from '@theia/ai-core/lib/common';
-import { FileContentFunction, GetWorkspaceFileList, FileCreateFunction } from './function';
-
-
-export default new ContainerModule(bind => {
-    // Replace this line with the desired binding, e.g. "bind(CommandContribution).to(TheiaAiExerciserContribution)
-    bind(CodingExerciseAgent).toSelf().inSingletonScope;
-    //bind(FileService).toSelf().inSingletonScope();
-    //bind(WorkspaceService).toSelf().inSingletonScope();
-    //bind(TerminalService).toSelf().inSingletonScope();
-    bind(ToolProvider).to(GetWorkspaceFileList);
-    bind(ToolProvider).to(FileContentFunction);
-    bind(ToolProvider).to(FileCreateFunction);
-    bind(Agent).toService(CodingExerciseAgent);
-    bind(ChatAgent).toService(CodingExerciseAgent);
-});
