--- theia-ai-exerciser/src/browser/function-name.ts
@@ -1,3 +1,0 @@
-export const FILE_CONTENT_FUNCTION_ID = 'getFileContent';
-export const GET_WORKSPACE_FILE_LIST_FUNCTION_ID = 'getWorkspaceFileList';
-export const CREATE_FILE_FUNCTION_ID = 'createFile';
